# path list optimizations