--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -8,7 +8,9 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Enhancements
 
-* None.  
+* PathList optimizations related to file system reads.  
+  [manuyavuz](https://github.com/manuyavuz)
+  [#issue_number](https://github.com/CocoaPods/CocoaPods/issues/4927)
 
 ##### Bug Fixes