--- a/lib/cocoapods/sandbox/path_list.rb
+++ b/lib/cocoapods/sandbox/path_list.rb
@@ -51,12 +51,14 @@ module Pod
         unless root.exist?
           raise Informative, "Attempt to read non existent folder `#{root}`."
         end
-
         dirs = []
         files = []
         root_length = root.cleanpath.to_s.length + File::SEPARATOR.length
-        Find.find(root.to_s) do |f|
+        Dir.glob(root + '**/*', File::FNM_DOTMATCH).each do |f|
           directory = File.directory?(f)
+          # Ignore `.` and `..` directories
+          next if directory && f =~ /\.\.?$/
+
           f = f.slice(root_length, f.length - root_length)
           next if f.nil?