- Uses Dir.glob instead of Find.find for faster content fetch
- Does not recalculate file system content for local pods during installation as they won't be changed by cocoapods itself anytime.


## Improvements
### Previous
```
bundle exec pod install --clean-install  10.97s user 7.97s system 96% cpu 19.628 total
bundle exec pod install  4.91s user 4.41s system 95% cpu 9.772 total
```

### This branch
```
bundle exec pod install --clean-install  6.62s user 4.12s system 86% cpu 12.453 total
bundle exec pod install  3.99s user 3.60s system 88% cpu 8.587 total
```