@manuyavuz 2019-12-23
Partially fixes #9427.