@dnkoutso 2019-12-24
Yeah we can add a changelog entry for "Various performance optimizations around file system reads" or something like that so we can know it went through.