@dnkoutso 2019-12-24
LGTM. Any tests to add here?