@manuyavuz 2019-12-24
@dnkoutso rebased onto `1-9-stable`. What is our branch policy, will you cherry pick into master after merge?