@manuyavuz 2019-12-24
For tests, there are existing tests for escaped path logic, but there is no test for `refresh_file_accessors` logic. There is no test for not cleaning local pods inside `pod_dir_cleaner` for example. 

If we want to have better coverage here, I can spend some time tonight maybe. But I'm not sure if this should go into this PR.